Proyecto_Retail_IA_GitHub/requirements.txt
# ===========================================================
# Dependencias del proyecto Retail Shelf CV v2.0
# ===========================================================

# --- Backend API ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# --- ML / Deep Learning ---
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# --- ML clasico ---
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0

# --- Ensemble / Fusion ---
ensemble-boxes>=1.0.9

# --- Procesamiento de imagenes ---
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# --- Configuracion ---
pyyaml>=6.0

# --- Utilidades ---
matplotlib>=3.7.0

# --- Opcional: Clustering avanzado ---
# hdbscan>=0.8.33
